requirements.txt
pandas
pandasql
python-can
python-telegram-bot
termcolor
rich
pyfiglet==0.7